Removing deprecated management commands which intended to clear historical data
