+++ openedx/core/djangoapps/util/row_delete.py
@@ -1,5 +1,32 @@
 """
 Code to delete rows from a table within a Django mgmt command using best practices.
+Following lines show how to use delete_rows():
+
+# Command to delete all rows from the student_historicalcourseenrollment table.
+
+
+import logging
+
+from openedx.core.djangoapps.util.row_delete import BaseDeletionCommand, delete_rows
+from student.models import CourseEnrollment
+
+log = logging.getLogger(__name__)
+
+
+class Command(BaseDeletionCommand):
+    # Example usage: ./manage.py lms --settings=devstack delete_historical_enrollment_data
+    help = 'Deletes all historical CourseEnrollment rows (in chunks).'
+
+    def handle(self, *args, **options):
+        # Deletes rows, chunking the deletes to avoid long table/row locks.
+        chunk_size, sleep_between = super(Command, self).handle(*args, **options)
+        delete_rows(
+            CourseEnrollment.objects,
+            'student_historicalcourseenrollment',
+            'history_id',
+            chunk_size, sleep_between
+        )
+
 """
